requirements.txt
flask
flask_session
flask_sqlalchemy
flask_login
flask_wtf
flask_bcrypt
pymysql
python-dotenv

openai